pyproject.toml
[project]
name = "hand-tracking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mediapipe==0.10.14",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pyqtgraph>=0.14.0",
    "pyside6-essentials>=6.10.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
